add listeners/mentionListener.js
